# api/openapi.yaml
openapi: 3.0.3
info:
  title: go-fiber-starter API
  description: Go + Fiber v3 + GORM 기반 REST API 보일러플레이트. /api/v1 업무 API의 JSON 성공·오류 응답은
    공통 Envelope를 사용한다. 204는 빈 본문이며 health 프로브 JSON, Prometheus 텍스트, OpenAPI YAML은 별도 형식이다.
  version: 1.0.0
servers:
- url: /
paths:
...
  /api/v1/tasks:
    post:
      summary: Task 생성
      operationId: createTask
      tags:
      - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              title: write tests
              due_date: '2030-01-02T03:04:05Z'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEnvelope'
              example:
                success: true
                data:
                  id: 1
                  title: write tests
                  done: false
                  due_date: '2030-01-02T03:04:05Z'
                  created_at: '2026-08-23T00:00:00+09:00'
                  updated_at: '2026-08-23T00:00:00+09:00'
        '400':
          description: 본문 파싱 실패(malformed request body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 필드 검증 실패(INVALID_REQUEST, details에 필드별 사유)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: INVALID_REQUEST
                  message: validation failed
                  details:
                  - field: Title
                    reason: required
        '429':
          description: 전역 rate limit 초과(RATE_LIMITED, IP별 RATE_LIMIT_PER_MINUTE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패(UNAUTHORIZED) — AUTH_ENABLED=true일 때만 발생
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - BearerAuth: []
    get:
      summary: Task 목록(offset 페이지네이션)
      operationId: listTasks
      tags:
      - tasks
      parameters:
      - name: page
        in: query
        required: false
        description: 페이지 번호(1 이상, 기본 1)
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: limit
        in: query
        required: false
        description: 페이지 크기(1~100, 기본 20, 100 초과분은 클램프)
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: 목록과 페이지 메타
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListEnvelope'
        '429':
          description: 전역 rate limit 초과(RATE_LIMITED, IP별 RATE_LIMIT_PER_MINUTE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패(UNAUTHORIZED) — AUTH_ENABLED=true일 때만 발생
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: POST /api/v1/auth/login으로 발급. AUTH_ENABLED=true일 때 /api/v1/tasks에
        요구된다.
...
  schemas:
...
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 안정적인 에러 식별자. 예 NOT_FOUND, INVALID_REQUEST, TASK_NOT_FOUND,
            INVALID_CREDENTIALS, RATE_LIMITED
        message:
          type: string
        details:
          type: array
          description: 필드별 상세(주로 422 검증 실패). validator ValidationErrors의 Field/ActualTag에서
            온다.
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
      - field
      - reason
      properties:
        field:
          type: string
        reason:
          type: string
          description: 위반한 validate 태그(예 required, min, max)
    TaskEnvelope:
      type: object
      required:
      - success
      - data
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/TaskResponse'
    TaskListEnvelope:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    ErrorEnvelope:
      type: object
      required:
      - success
      - error
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          $ref: '#/components/schemas/Error'
    CreateTaskRequest:
      type: object
      required:
      - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        due_date:
          type: string
          format: date-time
          nullable: true
...
    TaskResponse:
      type: object
      required:
      - id
      - title
      - done
      - created_at
      - updated_at
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        done:
          type: boolean
        due_date:
          type: string
          format: date-time
          nullable: true
          description: 마감일. 미설정 시 필드가 생략된다(omitempty).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PageMeta:
      type: object
      required:
      - page
      - limit
      - total_count
      - total_pages
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_count:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64
...